move boa.verify to verifiers
